Some brainstorming, and genderless robots
A bit more brainstorming in the comments about what to do next maybe

Also some minor changes to choosing your gender and it being different for robots (robots get to be genderless and get called “it” since robots aren’t people but machines)

And I guess some more flavor text about it being rude to talk about someone in the third person, sure, why not. Gotta keep this thing as a parody and not have it be too serious.